// examples/shared/src/schemas/auth.gen.ts
export const userModel = z.object({
  id: z.string(),
  name: z.string(),
  email: z.string(),
  emailVerified: z.boolean(),
  image: z.string(),
  createdAt: z.string(),
  updatedAt: z.string(),
})
export const sessionModel = z.object({
  id: z.string(),
  expiresAt: z.string(),
  token: z.string(),
  createdAt: z.string(),
  updatedAt: z.string(),
  ipAddress: z.string(),
  userAgent: z.string(),
  userId: z.string(),
})
// ...
export const getGetSession = defineGet('/get-session', {
  response: z.object({ session: sessionModel, user: userModel }),
})

export const postSignOut = definePost('/sign-out', {
  input: { json: z.object({}) },
  response: z.object({ success: z.boolean() }),
})

export const postSignUpEmail = definePost('/sign-up/email', {
  input: {
    json: z.object({
      name: z.string(),
      email: z.string(),
      password: z.string(),
      image: z.string(),
      callbackURL: z.string(),
      rememberMe: z.boolean(),
    }),
  },
  response: z.object({
    token: z.string().nullable(),
    user: z.object({
      id: z.string(),
      email: z.email(),
      name: z.string(),
      image: z.url().nullable(),
      emailVerified: z.boolean(),
      createdAt: z.iso.datetime(),
      updatedAt: z.iso.datetime(),
    }),
  }),
})

export const postSignInEmail = definePost('/sign-in/email', {
  input: {
    json: z.object({
      email: z.string(),
      password: z.string(),
      callbackURL: z.string().nullable(),
      rememberMe: z.string().nullable(),
    }),
  },
  response: z.object({
    redirect: z.boolean(),
    token: z.string(),
    url: z.object({}).nullable(),
    user: z.object({
      id: z.string(),
      email: z.string(),
      name: z.string().nullable(),
      image: z.string().nullable(),
      emailVerified: z.boolean(),
      createdAt: z.iso.datetime(),
      updatedAt: z.iso.datetime(),
    }),
  }),
})